Warning fixes.
2005-04-24  Eric Anholt <eta@lclark.edu>

        * kaffe/kaffevm/exception.c,
        kaffe/kaffevm/locks.c,
        kaffe/kaffevm/locks.h,
        kaffe/kaffevm/string.c,
        kaffe/kaffevm/intrp/checks.h,
        kaffe/kaffevm/intrp/icode.h,
        kaffe/kaffevm/jni/jni-base.c,
        kaffe/kaffevm/jni/jni-helpers.c,
        kaffe/kaffevm/jni/jni.c
        kaffe/kaffevm/jni/jni_funcs.h
        kaffe/kaffevm/kaffe-gc/gc-mem.c:
        Fixed a few warnings.

+++ kaffe/kaffevm/jni/jni-helpers.c
@@ -80,7 +80,7 @@ void *KaffeJNI_GetDirectBufferAddress(JNIEnv *env, jobject buffer)
   return address;
 }
 
-jint KaffeJNI_GetDirectBufferCapacity(JNIEnv *env UNUSED, jobject buffer)
+long KaffeJNI_GetDirectBufferCapacity(JNIEnv *env UNUSED, jobject buffer)
 {  
   jmethodID capacity_method;
   jint capacity;
